Add static values to be updaed for self introspection of /dev/mem
This is a bad bad commit. Please don't use this. I am doing this
just to keep a track of the values that needs to be changed to perform
self memory introspection using libvmi. You configure libvmi.conf to
read /dev/mem instead of a memory snapshot or guest memory.

To be able to do this, you will also have to custom-compile your kernel
with strict memory access of /dev/mem disabled. You don't want to do
this generally in any production machine.